requirements.txt
asgiref==3.3.1
Django==3.1.6
django-pgcrypto-fields==2.6.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
